backup_v2_20250814_011809/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
email-validator==2.2.0
Flask-WTF==1.2.1
